ok build appears to be all working, whew
