+++ src/cli/index.ts
@@ -1,2 +1,3 @@
 export { parseFlags } from './parse-flags';
 export { run, runTask } from './run';
+export { ConfigFlags } from './config-flags';
